De-duplicate webpack configs
