+++ config/webpack/modules.js
